+++ pollbot/telegram/callback_handler/vote.py
@@ -114,7 +114,7 @@ def handle_vote(session, context):
 
     # Additional test commit. Sometimes deadlocks happen right here.
     session.commit()
-    increase_user_stat(session, poll.user, "poll_callback_calls")
+    increase_user_stat(session, poll.user, "votes")
 
 
 def respond_to_vote(session, line, context, poll, remaining_votes=None, limited=False):